Fixd unit tests. One TODO, consider changing default weight to 1 instead of 0. See note in calculateAggregates unit test to see why.
