+++ lib/aggregate/calculate-aggregates.js
@@ -1,66 +1,16 @@
 const path = require('path');
 const Report = require('../selectors/report');
 const Method = require('../selectors/method');
-const cli = require('../cli/');
-
-function add(initialValue = 0) {
-  let value = initialValue;
-  return function (increment = 0) {
-    value += increment;
-    return value;
-  }
-}
-
-function weightedAverage() {
-  const denominator = add();
-  const numerator = add();
-  let average = 0;
-  return function(value = 0, weight = 0) {
-    if (weight) {
-      average = numerator(value * weight) / denominator(weight);
-    }
-    return average;
-  }
-}
-
-function makeMetricsObject() {
-  return {
-    cyclomatic: add(),
-    sloc: add(),
-    numberOfMethods: add(),
-    halstead: {
-      bugs: add(),
-      difficulty: add(),
-      volume: add(),
-    }
-  };
-}
-
-function makeAverage() {
-  return {
-    cyclomatic: weightedAverage(),
-    sloc: weightedAverage(),
-    halstead: {
-      bugs: weightedAverage(),
-      difficulty: weightedAverage(),
-      volume: weightedAverage(),
-    }
-  };
-}
-
-function withoutExtension(str) {
-  return str.replace(/(.jsx?|.ts)?$/, '');
-}
-
-const getAbsoluteDependencyPath = function (projectRoot, sourceFilePath, dependencyPath) {
-  if (dependencyPath.charAt(0) === '.') {
-    const sourceFileDirectory = path.dirname(sourceFilePath);
-    const res = path.resolve(sourceFileDirectory, dependencyPath);
-    return withoutExtension(path.relative(projectRoot, res));
-  }
-
-  return withoutExtension(dependencyPath);
-};
+const cli = require('../cli');
+const {
+  evaluate,
+  makeAverageObject,
+  makeMetricsObject,
+} = require('./model');
+const {
+  getAbsoluteDependencyPath,
+  withoutExtension,
+} = require('./utils');
 
 function getWeights(reports) {
   const weights = {};
@@ -93,29 +43,10 @@ function getReportTotals(report) {
   return total;
 }
 
-function evaluate(obj) {
-  if (typeof obj === 'function') {
-    return obj();
-  }
-
-  if (Array.isArray(obj)) {
-    return obj.forEach(evaluate);
-  }
-
-  if (typeof obj === 'object') {
-    return Object.keys(obj).reduce(
-      (acc, key) => ({ ...acc, [key]: evaluate(obj[key]) }),
-      {}
-    );
-  }
-
-  return obj;
-}
-
 function calculateAggregates(reports = []) {
   const total = makeMetricsObject();
-  const average = makeAverage();
-  const weightedAverage = makeAverage();
+  const average = makeAverageObject();
+  const weightedAverage = makeAverageObject();
   const weights = getWeights(reports);
   const details = {};
 
@@ -158,8 +89,7 @@ function calculateAggregates(reports = []) {
 }
 
 module.exports = {
-  add,
+  getWeights,
+  getReportTotals,
   calculateAggregates,
-  makeMetricsObject,
-  weightedAverage,
 };
