+++ lib/aggregate/index.js
@@ -21,11 +21,14 @@ async function aggregate(reports) {
     gitCommit = await getCommit(getSourceDirectoryPath());
     logger.debug(`git commit is ${gitCommit}`);
   } catch(err) {
-    logger.warn('Error obtaining git hash.', err);
-    if (err.type === 'NotARepoError') {
-      gitCommit = 'not a git repo';
-    } else {
-      gitCommit = 'error obtaining git commit hash';
+    /* istanbul ignore next */
+    {
+      logger.warn('Error obtaining git hash.', err);
+      if (err.type === 'NotARepoError') {
+        gitCommit = 'not a git repo';
+      } else {
+        gitCommit = 'error obtaining git commit hash';
+      }
     }
   }
