+++ lib/aggregate/utils.js
@@ -1,0 +1,20 @@
+const path = require('path');
+
+function withoutExtension(str) {
+  return str.replace(/(.jsx?|.ts)?$/, '');
+}
+
+const getAbsoluteDependencyPath = function (projectRoot, sourceFilePath, dependencyPath) {
+  if (dependencyPath.charAt(0) === '.') {
+    const sourceFileDirectory = path.dirname(sourceFilePath);
+    const res = path.resolve(sourceFileDirectory, dependencyPath);
+    return withoutExtension(path.relative(projectRoot, res));
+  }
+
+  return withoutExtension(dependencyPath);
+};
+
+module.exports = {
+  getAbsoluteDependencyPath,
+  withoutExtension,
+};
